LPD-44396 Reuse code
