+++ modules/apps/captcha/captcha-test/src/testIntegration/java/com/liferay/captcha/internal/function/captcha/test/FunctionCaptchaImplTest.java
@@ -6,6 +6,7 @@
 package com.liferay.captcha.internal.function.captcha.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.captcha.BaseCaptchaTestCase;
 import com.liferay.captcha.configuration.CaptchaConfiguration;
 import com.liferay.client.extension.type.CustomElementCET;
 import com.liferay.client.extension.type.configuration.CETConfiguration;
@@ -14,32 +15,16 @@ import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.configuration.test.util.CompanyConfigurationTemporarySwapper;
 import com.liferay.portal.configuration.test.util.ConfigurationTestUtil;
-import com.liferay.portal.kernel.portlet.PortletURLFactoryUtil;
-import com.liferay.portal.kernel.portlet.url.builder.PortletURLBuilder;
-import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
-import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-
-import java.net.HttpURLConnection;
-import java.net.URL;
-
 import java.util.Map;
-
-import javax.portlet.PortletMode;
-import javax.portlet.PortletRequest;
-import javax.portlet.WindowState;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -49,13 +34,11 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.springframework.mock.web.MockHttpServletRequest;
-
 /**
  * @author Pedro Victor Silvestre
  */
 @RunWith(Arquillian.class)
-public class FunctionCaptchaImplTest {
+public class FunctionCaptchaImplTest extends BaseCaptchaTestCase {
 
 	@ClassRule
 	@Rule
@@ -134,69 +117,10 @@ public class FunctionCaptchaImplTest {
 								"FunctionCaptchaImpl#" + servicePid
 						).build())) {
 
-			Assert.assertTrue(_isCaptchaRendered());
+			Assert.assertTrue(
+				isCaptchaRendered(
+					StringPool.LESS_THAN + _cet.getHTMLElementName()));
 		}
-	}
-
-	private MockHttpServletRequest _getMockHttpServletRequest()
-		throws Exception {
-
-		MockHttpServletRequest mockHttpServletRequest =
-			new MockHttpServletRequest();
-
-		ThemeDisplay themeDisplay = new ThemeDisplay();
-
-		themeDisplay.setLayout(
-			_layoutLocalService.fetchLayout(TestPropsValues.getPlid()));
-		themeDisplay.setPlid(TestPropsValues.getPlid());
-		themeDisplay.setPortalURL("http://localhost:8080");
-		themeDisplay.setScopeGroupId(TestPropsValues.getGroupId());
-		themeDisplay.setSiteGroupId(TestPropsValues.getGroupId());
-
-		mockHttpServletRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, themeDisplay);
-
-		return mockHttpServletRequest;
-	}
-
-	private boolean _isCaptchaRendered() throws Exception {
-		URL url = new URL(
-			PortletURLBuilder.create(
-				PortletURLFactoryUtil.create(
-					_getMockHttpServletRequest(), PortletKeys.LOGIN,
-					_layoutLocalService.fetchLayout(TestPropsValues.getPlid()),
-					PortletRequest.RENDER_PHASE)
-			).setMVCRenderCommandName(
-				"/login/create_account"
-			).setParameter(
-				"saveLastPath", false
-			).setPortletMode(
-				PortletMode.VIEW
-			).setWindowState(
-				WindowState.MAXIMIZED
-			).buildString());
-
-		HttpURLConnection httpURLConnection =
-			(HttpURLConnection)url.openConnection();
-
-		Assert.assertEquals(
-			HttpURLConnection.HTTP_OK, httpURLConnection.getResponseCode());
-
-		try (BufferedReader bufferedReader = new BufferedReader(
-				new InputStreamReader(httpURLConnection.getInputStream()))) {
-
-			String line;
-
-			while ((line = bufferedReader.readLine()) != null) {
-				if (line.contains(
-						StringPool.LESS_THAN + _cet.getHTMLElementName())) {
-
-					return true;
-				}
-			}
-		}
-
-		return false;
 	}
 
 	private static final String _VIRTUAL_HOSTNAME =
@@ -209,8 +133,5 @@ public class FunctionCaptchaImplTest {
 
 	private static String _pid;
 	private static Map<String, Object> _properties;
-
-	@Inject
-	private LayoutLocalService _layoutLocalService;
 
 }
